requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
behave==1.2.6
behave2cucumber==1.0.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.1
decorator==5.1.1
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
iso8601==1.0.2
jsonschema==4.4.0
more-itertools==8.12.0
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
packaging==21.3
paramiko==2.7.2
parse==1.19.0
parse-type==0.6.0
pluggy==0.13.1
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pycryptodome==3.14.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==4.6.3
PyYAML==5.3.1
requests==2.25.0
requests-toolbelt==0.9.1
retry==0.9.2
rpmfile==1.0.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scp==0.14.4
simplejson==3.17.2
six==1.16.0
soupsieve==2.3.2.post1
sshtunnel==0.4.0
tabulate==0.8.9
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.8.0
